backend/requirements/requiremente.txt
apturl==0.5.2
asgiref==3.3.1
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==3.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.5.0
distro-info===0.23ubuntu1
Django==3.1.5
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.2
docker==4.1.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
fasteners==0.14.1
future==0.18.2
gyp==0.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
jeepney==0.4.3
jsonschema==3.2.0
kazam==1.4.5
keyring==21.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.14.0
macaroonbakery==1.3.1
Mako==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
paramiko==2.7.1
pexpect==4.6.0
Pillow==7.2.0
protobuf==3.12.3
psycopg2-binary==2.8.6
pycairo==1.16.2
pycups==2.0.1
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.1.3+ubuntu1.3
python-dateutil==2.8.1
python-debian==0.1.37
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.47
requests==2.23.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==3.1.2
simplejson==3.17.0
six==1.15.0
sqlparse==0.4.1
systemd-python==234
texttable==1.6.2
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
wadllib==1.3.4
websocket-client==0.53.0
xkit==0.0.0
zipp==1.0.0
